+++ src/refactorings/lift-up-conditional/lift-up-conditional.test.ts
@@ -1,5 +1,5 @@
-import { ErrorReason, Code } from "../../editor/editor";
 import { InMemoryEditor } from "../../editor/adapters/in-memory-editor";
+import { Code, ErrorReason } from "../../editor/editor";
 import { testEach } from "../../tests-helpers";
 
 import { liftUpConditional } from "./lift-up-conditional";
@@ -37,7 +37,10 @@ describe("Lift Up Conditional", () => {
     doAnotherThing();
   }
 } else {
-  doAnotherThing();
+  if (isValid)
+    {} else {
+    doAnotherThing();
+  }
 }`
       },
       {